fixed scrolling and show real words
