+++ frontend/src/routes/App.tsx
@@ -16,8 +16,6 @@ function prettyPrintBalance(b: MutinyBalance): string {
 }
 
 function App() {
-  const [wasmSupported, setWasmSupported] = useState(true)
-
   const nodeManager = useContext(NodeManagerContext);
 
   const queryClient = useQueryClient()
@@ -45,39 +43,17 @@ function App() {
     queryClient.invalidateQueries({ queryKey: ['balance'] })
   }
 
-  useEffect(() => {
-    // https://stackoverflow.com/questions/47879864/how-can-i-check-if-a-browser-supports-webassembly
-    const checkWasm = async () => {
-      try {
-        if (typeof WebAssembly === "object"
-          && typeof WebAssembly.instantiate === "function") {
-          const module = new WebAssembly.Module(Uint8Array.of(0x0, 0x61, 0x73, 0x6d, 0x01, 0x00, 0x00, 0x00));
-          if (!(module instanceof WebAssembly.Module)) {
-            throw new Error("Couldn't instantiate WASM Module")
-          }
-        } else {
-          throw new Error("No WebAssembly global object found")
-        }
-      } catch (e) {
-        console.error(e)
-        setWasmSupported(false);
-      }
-    }
-    checkWasm();
-  }, [])
+  async function handleSync() {
+    await nodeManager?.sync()
+    queryClient.invalidateQueries({ queryKey: ['balance'] })
+  }
 
   return (
-    <div className="flex flex-col h-full w-full">
+    <>
       <header className='p-8'>
         <img src={logo} className="App-logo" alt="logo" />
         <h2>You're probably looking for <a href="/tests">the tests</a></h2>
-        <p>View the <a href="https://github.com/BitcoinDevShop/mutiny-web-poc">source</a></p>
-        {wasmSupported ? <p>WASM works!</p> :
-          <p>
-            WASM does not seem supported in your browser, this might not work for you!
-            You may have to turn on Javascript JIT in your browser settings.
-          </p>
-        }
+
       </header>
       <ScreenMain>
         {nodeManager ?
@@ -87,6 +63,7 @@ function App() {
             <h1 className='text-4xl font-light uppercase'>{balance && prettyPrintBalance(balance).toString()} <span className='text-2xl'>sats</span></h1>
             <div />
             <div className='flex flex-col gap-2 items-start'>
+              <button onClick={handleSync}>Sync</button>
               <button onClick={handleCheckBalance}>Check balance</button>
               <button className='green-button' onClick={handleNavSend}>Send</button>
               {/* TODO if no funds can do deposit instead of receive */}
@@ -100,7 +77,7 @@ function App() {
           : <h1>Loading...</h1>}
         <MutinyToaster />
       </ScreenMain>
-    </div>
+    </>
   );
 }
