+++ frontend/src/routes/Send.tsx
@@ -30,7 +30,7 @@ function Send() {
     }
   }
   return (
-    <div className="flex flex-col h-full w-full">
+    <>
       <header className='p-8 flex justify-between items-center'>
         <PageTitle title="Send" theme="green" />
         <Close />
@@ -49,7 +49,7 @@ function Send() {
         </div>
       </ScreenMain>
       <MutinyToaster />
-    </div>
+    </>
   );
 }
