+++ frontend/src/routes/Transactions.tsx
@@ -16,7 +16,7 @@ function SingleTransaction() {
 
 function Transactions() {
     return (
-        <div className="h-full">
+        <>
             <header className='px-8 pt-8 flex justify-between items-center'>
                 <PageTitle title="Transactions" theme="green" />
                 <Close />
@@ -33,7 +33,7 @@ function Transactions() {
                     <SingleTransaction />
                 </ul>
             </ScreenMain>
-        </div>
+        </>
     )
 }
